Create CalendarService.js
supports data interactions for CalendarCrud.js, specially backend REST calls. I guess most of functions are self-descripting; but only about getCalendarsWithFiltersAndSort(), this function first tranforms the lazyParams to the Object format which is supported by my back-end. so, obviously you need to have backend which exposes a search by Filters and Sort. and then you can create a new object which is compatible by your object in backend. and then just send it via a POST method to your backend. POST had to be used as you cannot attach a payload in your request body of a GET method. I mean by tools like PostMan it's possible. but I guess according to REST standard it is still controversial to do so.